+++ src/main/java/org/example/sivillage/brand/domain/BrandLike.java
@@ -1,4 +1,4 @@
-package org.example.sivillage.member.domain;
+package org.example.sivillage.brand.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
